+++ src/test/java/lotto/custom/service/LottoDrawingService/BonusNumberServiceTest.java
@@ -1,10 +1,11 @@
-package lotto.custom.service;
+package lotto.custom.service.LottoDrawingService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Arrays;
 import java.util.List;
+import lotto.custom.service.LottoDrawingService;
 import lotto.custom.validator.CustomErrorMessages;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
