+++ src/test/java/lotto/custom/service/LottoDrawingService/WinningNumberServiceTest.java
@@ -1,4 +1,4 @@
-package lotto.custom.service;
+package lotto.custom.service.LottoDrawingService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.Arrays;
 import java.util.List;
 import lotto.custom.common.ErrorMessages;
+import lotto.custom.service.LottoDrawingService;
 import lotto.custom.validator.CustomErrorMessages;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
